init_code/requirements.txt
asgiref==3.5.0
Django==4.0.3
django-appconf==1.0.5
django-ckeditor==6.2.0
django-cors-headers==3.11.0
django-imagekit==4.1.0
django-js-asset==2.0.0
django-rest-framework==0.1.0
django-templated-mail==1.1.1
djangorestframework==3.13.1
pilkit==2.0
Pillow==9.0.1
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
uuid==1.30
